server/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
Werkzeug==3.0.1
